+++ benches/analysis/mod.rs
@@ -3,7 +3,6 @@
 use std::ffi::OsStr;
 use std::path::Path;
 use std::path::PathBuf;
-use std::sync::Arc;
 
 use criterion::BenchmarkGroup;
 use criterion::Criterion;
@@ -53,8 +52,7 @@ impl AnalyzeWorkflows {
             let config = AnalysisConfig::default();
             let analyzer = Analyzer::new(
                 config,
-                Arc::new(wdl_modules::NullResolver),
-                None,
+                wdl::analysis::ResolutionContext::default(),
                 |_, _, _, _| async {},
             );
             analyzer.add_directory(&self.repo_root).await.unwrap();
@@ -71,8 +69,7 @@ impl AnalyzeWorkflows {
             let config = AnalysisConfig::default();
             let analyzer = Analyzer::new(
                 config,
-                Arc::new(wdl_modules::NullResolver),
-                None,
+                wdl::analysis::ResolutionContext::default(),
                 |_, _, _, _| async {},
             );
             let document = Url::from_file_path(self.repo_root.join(path)).unwrap();