+++ crates/wdl-analysis/tests/analysis.rs
@@ -18,7 +18,6 @@ use std::ffi::OsStr;
 use std::fs;
 use std::path::Path;
 use std::path::absolute;
-use std::sync::Arc;
 
 use anyhow::Context;
 use anyhow::Result;
@@ -167,8 +166,7 @@ async fn run_test(test: &Path) -> Result<(), anyhow::Error> {
     };
     let analyzer = Analyzer::new(
         config,
-        Arc::new(wdl_modules::NullResolver),
-        None,
+        wdl_analysis::ResolutionContext::default(),
         |_, _, _, _| async {},
     );