+++ crates/wdl-engine/tests/tasks.rs
@@ -22,7 +22,6 @@ use std::env;
 use std::fs;
 use std::path::Path;
 use std::path::absolute;
-use std::sync::Arc;
 use std::sync::LazyLock;
 
 use anyhow::Context;
@@ -70,8 +69,7 @@ fn run_test(test: &Path, config: TestConfig) -> BoxFuture<'_, Result<()>> {
 
         let analyzer = Analyzer::new(
             config.analysis,
-            Arc::new(wdl_modules::NullResolver),
-            None,
+            wdl_analysis::ResolutionContext::default(),
             |(), _, _, _| async {},
         );
         analyzer