+++ tests/lint_examples.rs
@@ -141,8 +141,7 @@ impl TestContext {
 
         let analyzer = Analyzer::new_with_validator(
             AnalysisConfig::default(),
-            std::sync::Arc::new(wdl_modules::NullResolver),
-            None,
+            wdl::analysis::ResolutionContext::default(),
             |_, _, _, _| async {},
             validator,
         );
@@ -164,8 +163,7 @@ impl TestContext {
             AnalysisConfig::default()
                 .with_diagnostics_config(DiagnosticsConfig::new(std::iter::once(rule)))
                 .with_fallback_version(Some(FALLBACK_VERSION)),
-            std::sync::Arc::new(wdl_modules::NullResolver),
-            None,
+            wdl::analysis::ResolutionContext::default(),
             |_, _, _, _| async {},
             validator,
         );